big re-factor done

+++ src/main/java/com/searchbox/solr/SenseLikeThisHandler.java
@@ -163,7 +163,7 @@ public class SenseLikeThisHandler extends RequestHandlerBase {
 
         DocList subFiltered = qr.getSubSetToSearchIn(filters);
         System.out.println("Number of documents to search:\t" + subFiltered.size());
-        slt = new SenseQuery(rtv, senseField, params.getDouble(SenseParams.SENSE_WEIGHT, SenseParams.DEFAULT_SENSE_WEIGHT), null);
+        slt = new SenseQuery(rtv, senseField, params.getFloat(SenseParams.SENSE_WEIGHT, SenseParams.DEFAULT_SENSE_WEIGHT), null);
         sltDocs = searcher.getDocListAndSet(slt, subFiltered, Sort.RELEVANCE, start, rows, flags);
 
         if (sltDocs == null) {
